requirements.txt
affine==2.3.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contextily==1.2.0
cycler==0.11.0
Fiona==1.8.21
fonttools==4.33.3
geographiclib==1.52
geopandas==0.11.0
geopy==2.2.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
mercantile==1.2.1
munch==2.5.0
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pyparsing==3.0.9
pyproj==3.3.1
pyTelegramBotAPI==4.6.0
python-dateutil==2.8.2
pytz==2022.1
rasterio==1.2.10
requests==2.28.1
scipy==1.8.1
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
snuggs==1.4.7
urllib3==1.26.9
xyzservices==2022.6.0
